rename models/evaluators to modules/metrics
