+++ src/kwja/metrics/char.py
@@ -9,7 +9,7 @@ from torchmetrics import Metric
 
 from kwja.callbacks.utils import convert_predictions_into_tags, set_morphemes
 from kwja.datamodule.datasets import CharDataset
-from kwja.evaluators.utils import unique
+from kwja.metrics.utils import unique
 from kwja.utils.constants import IGNORE_INDEX, WORD_NORM_OP_TAGS
 from kwja.utils.sub_document import extract_target_sentences, to_orig_doc_id
