+++ src/kwja/metrics/typo.py
@@ -6,7 +6,7 @@ from torchmetrics import Metric
 
 from kwja.callbacks.utils import apply_edit_operations, convert_predictions_into_typo_corr_op_tags
 from kwja.datamodule.datasets import TypoDataset
-from kwja.evaluators.utils import unique
+from kwja.metrics.utils import unique
 
 
 class TypoModuleMetric(Metric):
